% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Create a Set of Circle Co-Ordinates}
\usage{
circle(x, y, r = 1)
}
\arguments{
\item{x}{x position of the center of the circle}

\item{y}{y position of the center of the circle}

\item{r}{radius of the circle}
}
\value{
Matrix of x,y co-ordinates for a circle
}
\description{
Creates a set of circle co-ordinates, of radius r, at position x,y
}
\examples{
plot(circle(0, 0, r = 1), type = 'l', asp = 1)
}
