% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Volcontour}
\alias{Volcontour}
\title{Re-formatted version of the base dataset volcano.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1941 rows and 3 columns.
}
\usage{
Volcontour
}
\description{
A dataset containg the x,y,z co-ordinates of the
base data set volcano. Data is cropped in a radius
 <25 from the center of the volcano crater
}
\details{
@format a data frame with 1941 obs. and 3 variables

 @source r base package
}
\keyword{datasets}
